% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_csv.R
\name{make_csv}
\alias{make_csv}
\title{make a csv file of metadata}
\usage{
make_csv(dir_path)
}
\arguments{
\item{dir_path}{path to a directory where the \code{csv} file will be saved}
}
\value{
a \code{\link[base]{data.frame}} of metadata will be returned
invisibly
}
\description{
After running \code{\link[curatedMetagenomicDataPipeline]{save_rda}},
\code{make_csv} is used to create a \code{csv} file describing the \code{rda} files as
specified by \code{\link[ExperimentHubData]{makeExperimentHubMetadata}}.
Here, \code{dir_path} should be the date-specific directory (e.g.
\verb{~/uploads/2021-04-05/}) that was created by
\code{\link[curatedMetagenomicDataPipeline]{save_rda}}. That is, users should
provide a \code{dir_path} down to the date-level, but not further. When run, an
\code{ExperimentHub} ready metadata \code{csv} file will be saved at the location
specified.
}
\examples{
\dontrun{

make_csv("~/uploads/2021-04-05/")
}

}
\seealso{
\code{\link[=save_rda]{save_rda()}}
}
