% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dir.R
\name{read_dir}
\alias{read_dir}
\title{read a directory of tsv files}
\usage{
read_dir(dir_path)
}
\arguments{
\item{dir_path}{path to a directory of \code{tsv} files generated by the high load
pipeline}
}
\value{
A single-element named \code{\link[base]{list}} (see above) with a
\code{\link[Matrix]{dgTMatrix-class}} matrix or a \code{\link[base]{matrix}}.
}
\description{
Given a directory of \code{tsv} files, \code{read_dir()} will read in each file and
merge them into a single matrix. A \code{\link[Matrix]{dgTMatrix-class}}
matrix will be returned if the directory is named \code{genefamilies_relab} and a
\code{\link[base]{matrix}} will be returned in all other cases. In either
case, the matrix will be returned as a named \code{\link[base]{list}} element
with the correct name to be used when uploading to AWS S3.
}
