% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tar.R
\name{read_tar}
\alias{read_tar}
\title{read a tar file into a list}
\usage{
read_tar(tar_file)
}
\arguments{
\item{tar_file}{A path to a tar file generated by the high load pipeline.}
}
\value{
A \code{\link[base]{list}} of \code{\link[base]{matrix}} elements.
}
\description{
\code{read_tar()} will read in a \code{tar} file, zipped or not, and join \code{tsv} file(s)
in each directory. Internally, \code{tsv} file(s) are read in parallel using the
\code{\link{furrr}} package and \code{\link[readr]{read_tsv}}. When reading
from the \code{metaphlan_bugs_list} directory, the NCBI Taxonomy ID column is
skipped, and, if present, the \code{metadata} directory is ignored.
}
