% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_rda.R
\name{save_rda}
\alias{save_rda}
\title{save a list of matrix objects}
\usage{
save_rda(list_obj, dir_path)
}
\arguments{
\item{list_obj}{list object obtained from
\code{\link[curatedMetagenomicDataPipeline]{read_dir}}}

\item{dir_path}{path to a directory where the \code{rda} file will be saved}
}
\value{
the \code{\link[base]{NULL}} object will be returned invisibly
}
\description{
When the output of \code{\link[curatedMetagenomicDataPipeline]{read_dir}} is
piped to \code{save_rda} (and \code{dir_path} is provided), an \code{rda} file ready to be
uploaded to AWS S3 will be saved at the location specified. Here, \code{dir_path}
should be the top-level directory (e.g. \verb{~/uploads/}) in which an organized
directory structure will be created. That is, users should not provided a
nested directory path, existing after the first run, to \code{dir_path}.
}
\examples{
\dontrun{

read_dir("~/AsnicarF_2017/genefamilies_relab/") \%>\%
    save_rda("~/uploads/")
}

}
\seealso{
\code{\link[=read_dir]{read_dir()}}
}
